Converted `datasets.py` into sub-package
