+++ minerva/inbuilt_cfgs/example_3rd_party.yml
@@ -101,14 +101,14 @@ dataset_params:
                     module: minerva.transforms
 
             images_1:
-                module: minerva.datasets
+                module: minerva.datasets.__testing
                 name: TstImgDataset
                 root: NAIP
                 params:
                     res: 1.0
 
             image2:
-                module: minerva.datasets
+                module: minerva.datasets.__testing
                 name: TstImgDataset
                 root: NAIP
                 params:
@@ -118,7 +118,7 @@ dataset_params:
             transforms:
                 SingleLabel:
                     module: minerva.transforms
-            module: minerva.datasets
+            module: minerva.datasets.__testing
             name: TstMaskDataset
             root: Chesapeake7
             params:
@@ -138,7 +138,7 @@ dataset_params:
             transforms:
                 ToRGB:
                     module: minerva.transforms
-            module: minerva.datasets
+            module: minerva.datasets.__testing
             name: TstImgDataset
             root: NAIP
             params:
@@ -148,7 +148,7 @@ dataset_params:
             transforms:
                 SingleLabel:
                     module: minerva.transforms
-            module: minerva.datasets
+            module: minerva.datasets.__testing
             name: TstMaskDataset
             root: Chesapeake7
             params:
@@ -168,7 +168,7 @@ dataset_params:
             transforms:
                 ToRGB:
                     module: minerva.transforms
-            module: minerva.datasets
+            module: minerva.datasets.__testing
             name: TstImgDataset
             root: NAIP
             params:
@@ -178,7 +178,7 @@ dataset_params:
             transforms:
                 SingleLabel:
                     module: minerva.transforms
-            module: minerva.datasets
+            module: minerva.datasets.__testing
             name: TstMaskDataset
             root: Chesapeake7
             params:
