+++ minerva/inbuilt_cfgs/example_GeoSimConvNet.yml
@@ -126,13 +126,13 @@ dataset_params:
                     kernel_size: 25
 
             image1:
-                module: minerva.datasets
+                module: minerva.datasets.__testing
                 name: TstImgDataset
                 root: NAIP
                 params:
                     res: 1.0
             image2:
-                module: minerva.datasets
+                module: minerva.datasets.__testing
                 name: TstImgDataset
                 root: NAIP
                 params:
@@ -149,14 +149,14 @@ dataset_params:
                 length: 32
 
         image:
-            module: minerva.datasets
+            module: minerva.datasets.__testing
             name: TstImgDataset
             root: NAIP
             params:
                 res: 1.0
 
         mask:
-            module: minerva.datasets
+            module: minerva.datasets.__testing
             name: TstMaskDataset
             root: Chesapeake7
             params:
@@ -173,14 +173,14 @@ dataset_params:
                 length: 32
 
         image:
-            module: minerva.datasets
+            module: minerva.datasets.__testing
             name: TstImgDataset
             root: NAIP
             params:
                 res: 1.0
 
         mask:
-            module: minerva.datasets
+            module: minerva.datasets.__testing
             name: TstMaskDataset
             root: Chesapeake7
             params:
