+++ minerva/inbuilt_cfgs/example_UNetR_config.yml
@@ -100,13 +100,13 @@ dataset_params:
                 Normalise:
                     module: minerva.transforms
                     norm_value: 255
-            module: minerva.datasets
+            module: minerva.datasets.__testing
             name: TstImgDataset
             root: NAIP
             params:
                 res: 1.0
         mask:
-            module: minerva.datasets
+            module: minerva.datasets.__testing
             name: TstMaskDataset
             root: Chesapeake7
             params:
@@ -127,13 +127,13 @@ dataset_params:
                 Normalise:
                     module: minerva.transforms
                     norm_value: 255
-            module: minerva.datasets
+            module: minerva.datasets.__testing
             name: TstImgDataset
             root: NAIP
             params:
                 res: 1.0
         mask:
-            module: minerva.datasets
+            module: minerva.datasets.__testing
             name: TstMaskDataset
             root: Chesapeake7
             params:
@@ -153,13 +153,13 @@ dataset_params:
                 Normalise:
                     module: minerva.transforms
                     norm_value: 255
-            module: minerva.datasets
+            module: minerva.datasets.__testing
             name: TstImgDataset
             root: NAIP
             params:
                 res: 1.0
         mask:
-            module: minerva.datasets
+            module: minerva.datasets.__testing
             name: TstMaskDataset
             root: Chesapeake7
             params:
