+++ minerva/inbuilt_cfgs/example_autoencoder_config.yml
@@ -101,14 +101,14 @@ dataset_params:
                 Normalise:
                     module: minerva.transforms
                     norm_value: 255
-            module: minerva.datasets
+            module: minerva.datasets.__testing
             name: TstImgDataset
             root: NAIP
             params:
                 res: 1.0
 
         mask:
-            module: minerva.datasets
+            module: minerva.datasets.__testing
             name: TstMaskDataset
             root: Chesapeake7
             params:
@@ -129,14 +129,14 @@ dataset_params:
                 Normalise:
                     module: minerva.transforms
                     norm_value: 255
-            module: minerva.datasets
+            module: minerva.datasets.__testing
             name: TstImgDataset
             root: NAIP
             params:
                 res: 1.0
 
         mask:
-            module: minerva.datasets
+            module: minerva.datasets.__testing
             name: TstMaskDataset
             root: Chesapeake7
             params:
@@ -157,14 +157,14 @@ dataset_params:
                 Normalise:
                     module: minerva.transforms
                     norm_value: 255
-            module: minerva.datasets
+            module: minerva.datasets.__testing
             name: TstImgDataset
             root: NAIP
             params:
                 res: 1.0
 
         mask:
-            module: minerva.datasets
+            module: minerva.datasets.__testing
             name: TstMaskDataset
             root: Chesapeake7
             params:
